requirements.txt
numpy==1.24.3
cloudpickle==2.2.1
pickleshare==0.7.5
streamlit==1.23.1
streamlit-option-menu==0.3.6
scikit-learn==1.2.2
seaborn==0.12.2
pandas==2.0.2
matplotlib-base==3.7.1
matplotlib-inline==0.1.6



